Named pipes connection support
Added a very simple check to a checkbox. It relies on the checkbox setting the port to -1 inspired from Mysql documentation on the connectionstring to determine to use named pipes or tcp.

I also ran ctrl+a+k+f on all the files to clean up any coding format (compiled/ran fine so I would assume that I didn't harm anything)